requirements.txt
torch==2.1.2
transformers==4.37.2
tokenizers==0.15.0
peft==0.7.1
hydra-core==1.3.2
datasets==2.16.1
accelerate==0.25.0
flash-attn==2.3.6